* Added solar system implemetnation * Fixed accelerations calculation in model * Recalculated methods compration * Moved some class to files
