+++ view_model.py
@@ -3,102 +3,13 @@ import math
 from PyQt5 import QtCore, QtWidgets, QtGui
 import model
 import logging
+from viewModel.emitter import Emitter
+from viewModel.particle_processor import ParticleProcessor
+from viewModel.particle import Particle
 logging.basicConfig(format='%(asctime)s %(levelname)s:%(message)s', level=logging.DEBUG)
 from PyQt5.QtCore import (QObject, QPointF, 
         QPropertyAnimation, pyqtProperty)
 import queue
-
-class Arrow(QtWidgets.QGraphicsLineItem):
-    def __init__(self, x, y, radius, degree):
-        self.radius = radius
-        self.start = QtCore.QPointF(x, y)
-        rad = np.deg2rad(degree)
-        super().__init__(
-            x + radius * np.sin(rad),
-            y + radius * np.cos(rad),
-            x + 3 * radius * np.sin(rad),
-            y + 3 * radius * np.cos(rad)
-        )
-
-class Emitter(QtWidgets.QGraphicsEllipseItem):
-    def __init__(self, radius, degree, spawnInterval):
-        """
-        radius - emitter circle radius
-        degree - arrow rotation degree (0, 360)
-        spawnInterval - (a, b) interval. Between this interval emitter will
-        spawn particle
-        """
-        super().__init__(
-            QtCore.QRectF(-1 * radius, -1 * radius, 2 * radius, 2 * radius)
-        )
-        self.radius = radius
-        self.degree = degree
-        self.spawnInterval = spawnInterval
-        self.setFlag(QtWidgets.QGraphicsItem.ItemIsMovable, True)
-        self.arrow = Arrow(self.x(), self.y(), radius, degree)
-        self.arrow.setParentItem(self)
-        self.particleMass = 10000000
-        self.particle_speed_x = 0
-        self.particle_speed_y = 0
-        self.double_validator = QtGui.QDoubleValidator(
-            -10**9, 10**9, 3
-        )
-        self.particle_life_time = 10
-
-    def setVector(self, degree):
-        logging.info("Rotate emitter vector, degree: {0}".format(degree))
-        self.degree = degree
-        self.arrow.setRotation(degree)
-
-    def setParticleMass(self, mass):
-        if self.double_validator.validate(mass, 0)[0] == QtGui.QValidator.Acceptable:
-            self.particleMass = float(mass.replace(",", "."))
-
-    def setParticleSpeedX(self, speed_x):
-        if self.double_validator.validate(speed_x, 0)[0] == QtGui.QValidator.Acceptable:
-            self.particle_speed_x = float(speed_x.replace(",", "."))
-
-    def setParticleSpeedY(self, speed_y):
-        if self.double_validator.validate(speed_y, 0)[0] == QtGui.QValidator.Acceptable:
-            self.particle_speed_y = float(speed_y.replace(",", "."))
-
-    def setParticleLifeTime(self, life_time):
-        if self.double_validator.validate(life_time, 0)[0] == QtGui.QValidator.Acceptable:
-            self.particle_life_time = float(life_time.replace(",", "."))
-
-class Particle(QObject):
-    def __init__(self, radius, model):
-        super().__init__()
-        self.radius = radius
-        self.item = QtWidgets.QGraphicsEllipseItem(
-            QtCore.QRectF(-1 * radius, -1 * radius, 2 * radius, 2 * radius)
-        )
-        self.model = model
-        self._set_pos(QtCore.QPointF(model.x, model.y))
-    def _set_pos(self, pos):
-        self.item.setPos(pos)
-    pos = pyqtProperty(QPointF, fset=_set_pos)
-
-class ParticleProcessor(QtCore.QThread):
-    send_particles_signal = QtCore.pyqtSignal(list)
-    def __init__(self, solver, time_interval, q_in):
-        """
-        solver - model_solver
-        q_in - queue to get particles
-        time_interval - time interval for solver
-        """
-        super().__init__()
-        self.solver = solver
-        self.time_interval = time_interval
-        self.q_in = q_in
-
-    def run(self):
-        curr_particles = self.q_in.get()
-        while True:
-            if len(curr_particles) > 0:
-                result = self.solver.solve(curr_particles, self.time_interval)
-                self.send_particles_signal.emit(result)
-            curr_particles = self.q_in.get()
 
 class UniverseView(QtWidgets.QGraphicsView):
     def __init__(self, scene):
@@ -106,9 +17,10 @@ class UniverseView(QtWidgets.QGraphicsView):
         self.emitter = Emitter(10, 0, (10, 40))
         self.q_in_particles = queue.Queue()
         self.scene().addItem(self.emitter)
-        self.time_step = 0.1
-        self.time_interval = 1
+        self.time_step = 10**4
+        self.time_interval = 10**5
         self.solver = model.VerleSimpleSolver(time_step = self.time_step)
+        self.scale_factor = 1
         self._setup()
 
     def _setup(self):
@@ -126,7 +38,7 @@ class UniverseView(QtWidgets.QGraphicsView):
 
     @staticmethod
     def mass2radius(mass):
-        return math.log(float(mass), 6)
+        return mass
 
     def spawnParticle(self):
         """
@@ -175,9 +87,9 @@ class UniverseView(QtWidgets.QGraphicsView):
             )
         )
         logging.debug(
-            "First particles start pos:({0},{1}), end pos:({2},{3})".format(
-                np.round(particle_scenes[0][0].x,2), np.round(particle_scenes[0][0].y,2),
-                np.round(particle_scenes[-1][0].x,2), np.round(particle_scenes[-1][0].y,2)
+            "Start particles pos:{0},\n end particles pos:{1}".format(
+                [(np.round(p.x, 2), np.round(p.y, 2)) for p in particle_scenes[0]],
+                [(np.round(p.x, 2), np.round(p.y, 2)) for p in particle_scenes[-1]],
             )
         )
         for i, particles_scene in enumerate(particle_scenes[:-1]):
@@ -185,7 +97,7 @@ class UniverseView(QtWidgets.QGraphicsView):
             for j, v in enumerate(zip(particles_scene, self.particles)):
                 model, particle = v
                 particle_anim = QtCore.QPropertyAnimation(particle, b"pos")
-                particle_anim.setDuration(self.time_step * 10**3)
+                particle_anim.setDuration(self.time_step * 10 **-3)
                 particle_anim.setStartValue(QPointF(model.x, model.y))
                 particle_anim.setEndValue(
                     QPointF(
@@ -236,3 +148,145 @@ class UniverseView(QtWidgets.QGraphicsView):
         )
         self.particle_processor.start()
         self.particle_processor.send_particles_signal.connect(self.applyAnimation)
+
+    def addSolarSystem(self):
+        self.clear()
+        #sun
+        self.scale_factor = 9 * 10**-10
+        self.addParticle(
+            Particle(
+                radius = 2 * 10 ** 10,
+                model = model.Particle(
+                    x = 0,
+                    y = 0,
+                    mass = 1990000000000000000000000000000,
+                    speed_x=0,
+                    speed_y=0,
+                    life_time=np.inf,
+                )
+            )
+        )
+
+        #Merkury
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 58000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 48000,
+                    mass = 3.3e23,
+                )
+            )
+        )
+
+        #Venus
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 108000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 35000,
+                    mass = 4.87e24,
+                )
+            )
+        )
+
+        #Earth
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 150000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 30000,
+                    mass = 5.98e24,
+                )
+            )
+        )
+
+        #Mars
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 228000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 24000,
+                    mass = 6.42e23,
+                )
+            )
+        )
+
+        #Jupyter
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 778000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 13000,
+                    mass = 1.8999999999999998e27,
+                )
+            )
+        )
+
+        #Saturn
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 1427000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 9600,
+                    mass = 5.69e26,
+                )
+            )
+        )
+
+        #Uranus
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 2870000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 6800,
+                    mass = 8.68e25,
+                )
+            )
+        )
+
+        #Neptune
+        self.addParticle(
+            Particle(
+                radius = 0.5 * 10 ** 10,
+                model = model.Particle(
+                    x = 4497000000000.0,
+                    y = 0,
+                    life_time = np.inf,
+                    speed_x = 0,
+                    speed_y = 5400,
+                    mass = 1.0199999999999999e26,
+                )
+            )
+        )
+
+        self.scale(self.scale_factor, self.scale_factor)
+        self.emitter.setScale(1 / self.scale_factor)
+        self.emitter.setPos(0, 100)
